# Use vtproto Unmarshal and mempooling for zip.ManifiestEntry in buildbuddy_server.go