--- a/proto/BUILD
+++ b/proto/BUILD
@@ -92,6 +92,7 @@ proto_library(
         ":context_proto",
         ":remote_execution_proto",
         ":resource_proto",
+        "@com_github_planetscale_vtprotobuf//include/github.com/planetscale/vtprotobuf/vtproto:vtproto_proto",
         "@com_google_protobuf//:duration_proto",
         "@com_google_protobuf//:field_mask_proto",
         "@com_google_protobuf//:timestamp_proto",
@@ -635,6 +636,7 @@ proto_library(
     deps = [
         ":context_proto",
         ":remote_execution_proto",
+        "@com_github_planetscale_vtprotobuf//include/github.com/planetscale/vtprotobuf/vtproto:vtproto_proto",
     ],
 )
 
@@ -1326,6 +1328,7 @@ go_proto_library(
         ":context_go_proto",
         ":remote_execution_go_proto",
         ":resource_go_proto",
+        "@com_github_planetscale_vtprotobuf//include/github.com/planetscale/vtprotobuf/vtproto:vtproto_go_proto",
         "@org_golang_google_genproto_googleapis_rpc//status",
     ],
 )
@@ -1517,6 +1520,7 @@ go_proto_library(
     deps = [
         ":context_go_proto",
         ":remote_execution_go_proto",
+        "@com_github_planetscale_vtprotobuf//include/github.com/planetscale/vtprotobuf/vtproto:vtproto_go_proto",
     ],
 )
 
@@ -1921,6 +1925,12 @@ ts_proto_library(
 )
 
 ts_proto_library(
+    name = "vtproto_ts_proto",
+    proto = "@com_github_planetscale_vtprotobuf//include/github.com/planetscale/vtprotobuf/vtproto:vtproto_proto",
+    deps = ["descriptor_ts_proto"],
+)
+
+ts_proto_library(
     name = "cache_ts_proto",
     proto = ":cache_proto",
     deps = [
@@ -1931,6 +1941,7 @@ ts_proto_library(
         ":remote_execution_ts_proto",
         ":resource_ts_proto",
         ":timestamp_ts_proto",
+        ":vtproto_ts_proto",
     ],
 )
 
@@ -2036,5 +2047,8 @@ ts_proto_library(
 ts_proto_library(
     name = "zip_ts_proto",
     proto = ":zip_proto",
-    deps = [":context_ts_proto"],
+    deps = [
+        ":context_ts_proto",
+        ":vtproto_ts_proto",
+    ],
 )