<!-- Optional: Provide additional context (beyond the PR title). -->

<!-- Optional: link a GitHub issue.
     Example: "Fixes #123" will auto-close #123 when the PR is merged. -->

**Related issues**: N/A
ManifestEntry is a good candidate for mempooling since Unmarshal with the pool is 50% faster, compared to 34% faster without the pool. Also, we don't need to create a ManifestEntry for each http request. 
```
goos: linux
goarch: amd64
cpu: AMD Ryzen 9 7900X 12-Core Processor            
                                  │     Old      │                 New                 │                WithPool                │
                                  │    sec/op    │   sec/op     vs base                │    sec/op     vs base                  │
Marshal/pbName=ManifestEntry-24     388.3n ± 44%   242.0n ± 9%        ~ (p=0.105 n=10)
Unmarshal/pbName=ManifestEntry-24   546.9n ± 25%   383.6n ± 4%  -29.85% (p=0.002 n=10)   271.9n ± 35%  -50.29% (p=0.000 n=10)
geomean                             460.8n         304.7n       -33.87%                  271.9n        -50.29%                ¹
¹ benchmark set differs from baseline; geomeans may not be comparable

                                  │      Old      │                  New                  │                 WithPool                  │
                                  │      B/s      │      B/s       vs base                │      B/s        vs base                   │
Marshal/pbName=ManifestEntry-24     95.80Mi ± 80%   153.69Mi ± 8%        ~ (p=0.105 n=10)
Unmarshal/pbName=ManifestEntry-24   68.01Mi ± 34%    96.96Mi ± 4%  +42.56% (p=0.002 n=10)   136.82Mi ± 26%  +101.17% (p=0.000 n=10)
geomean                             80.72Mi          122.1Mi       +51.24%                   136.8Mi        +101.17%                ¹
¹ benchmark set differs from baseline; geomeans may not be comparable

                                  │     Old     │                 New                  │               WithPool               │
                                  │    B/op     │    B/op      vs base                 │    B/op     vs base                  │
Marshal/pbName=ManifestEntry-24      48.00 ± 0%    48.00 ± 0%       ~ (p=1.000 n=10) ¹
Unmarshal/pbName=ManifestEntry-24   128.00 ± 0%   128.00 ± 0%       ~ (p=1.000 n=10) ¹   32.00 ± 0%  -75.00% (p=0.000 n=10)
geomean                              78.38         78.38       +0.00%                    32.00       -75.00%                ²
¹ all samples are equal
² benchmark set differs from baseline; geomeans may not be comparable

                                  │    Old     │                 New                 │               WithPool               │
                                  │ allocs/op  │ allocs/op   vs base                 │ allocs/op   vs base                  │
Marshal/pbName=ManifestEntry-24     1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=10) ¹
Unmarshal/pbName=ManifestEntry-24   2.000 ± 0%   2.000 ± 0%       ~ (p=1.000 n=10) ¹   1.000 ± 0%  -50.00% (p=0.000 n=10)
geomean                             1.414        1.414       +0.00%                    1.000       -50.00%                ²
¹ all samples are equal
² benchmark set differs from baseline; geomeans may not be comparable
```